+++ module-member/src/main/java/com/gaethering/modulemember/exception/GlobalExceptionHandler.java
@@ -1,9 +1,11 @@
 package com.gaethering.modulemember.exception;
 
 import com.gaethering.modulecore.exception.ErrorResponse;
+import javax.mail.SendFailedException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mail.MailSendException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -20,6 +22,17 @@ public class GlobalExceptionHandler {
             .build();
 
         return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    @ExceptionHandler(MailSendException.class)
+    public ResponseEntity<ErrorResponse> handleMailSendException(MailSendException e) {
+
+        ErrorResponse response = ErrorResponse.builder()
+            .code(MemberErrorCode.FAILED_SEND_EMAIL.getCode())
+            .message(MemberErrorCode.FAILED_SEND_EMAIL.getMessage())
+            .build();
+
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
 
 }
