+++ module-member/src/main/java/com/gaethering/modulemember/service/EmailService.java
@@ -1,0 +1,9 @@
+package com.gaethering.modulemember.service;
+
+public interface EmailService {
+
+    void sendAuthMail(String email, String authCode);
+
+    void confirmAuthCode(String code);
+
+}
